parser/pageparser: Split the page lexer into some more files
See #5534
